openbts-python/requirements.txt
argparse==1.2.1
coverage==3.7.1
enum34==1.0.4
envoy==0.0.3
mock==1.0.1
nose==1.3.4
pyzmq==14.5.0
wsgiref==0.1.2
